fix previous upload problems
